minor cleanup

+++ contracts/staking/ValidatorManager.sol
@@ -124,7 +124,7 @@ abstract contract ValidatorManager is Initializable, ContextUpgradeable, IValida
         );
         uint256 numInitialValidators = subnetConversionData.initialValidators.length;
         require(
-            numInitialValidators < type(uint64).max, "ValidatorManager: too many initial validators"
+            numInitialValidators < type(uint32).max, "ValidatorManager: too many initial validators"
         );
 
         // Verify that the sha256 hash of the Subnet conversion data matches with the Warp message's subnetConversionID.
@@ -135,7 +135,7 @@ abstract contract ValidatorManager is Initializable, ContextUpgradeable, IValida
             subnetConversionData.validatorManagerAddress,
             uint32(numInitialValidators)
         );
-        for (uint64 i; i < numInitialValidators; i++) {
+        for (uint32 i; i < numInitialValidators; i++) {
             InitialValidator memory initialValidator = subnetConversionData.initialValidators[i];
             bytes32 nodeID = initialValidator.nodeID;
             require(nodeID != bytes32(0), "ValidatorManager: invalid node ID");
